Remove managed environment cache
The PAL already caches- no need to do it in managed.
Fold the code back into Environment.cs and switch over to
use Hashtable as per the legacy code.
